conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asttokens=2.4.1=pyhd8ed1ab_0
blas=1.0=mkl
brotli=1.0.9=h2bbff1b_8
brotli-bin=1.0.9=h2bbff1b_8
bzip2=1.0.8=hcfcfb64_5
ca-certificates=2024.6.2=h56e8100_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.0=py312h59b6b97_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.8.1=py312h53d5487_0
decorator=5.1.1=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
expat=2.6.2=h63175ca_0
fonttools=4.51.0=py312h2bbff1b_0
freetype=2.12.1=hdaf720e_2
glib=2.80.2=h7025463_1
glib-tools=2.80.2=h4394cf3_1
gst-plugins-base=1.24.4=hba88be7_0
gstreamer=1.24.4=h5006eae_0
icu=73.2=h63175ca_0
importlib-metadata=7.1.0=pyha770c72_0
importlib_metadata=7.1.0=hd8ed1ab_0
intel-openmp=2023.1.0=h59b6b97_46320
ipykernel=6.29.4=pyh4bbf305_0
ipython=8.25.0=pyh7428d3b_0
jedi=0.19.1=pyhd8ed1ab_0
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.7.2=py312h2e8e312_0
kiwisolver=1.4.4=py312hd77b12b_0
krb5=1.21.2=heb0366b_0
lcms2=2.16=h67d730c_0
lerc=4.0.0=h63175ca_0
libbrotlicommon=1.0.9=h2bbff1b_8
libbrotlidec=1.0.9=h2bbff1b_8
libbrotlienc=1.0.9=h2bbff1b_8
libclang13=18.1.7=default_h97ce8ae_0
libdeflate=1.20=hcfcfb64_0
libexpat=2.6.2=h63175ca_0
libffi=3.4.4=hd77b12b_1
libglib=2.80.2=h7025463_1
libiconv=1.17=hcfcfb64_2
libintl=0.22.5=h5728263_2
libintl-devel=0.22.5=h5728263_2
libjpeg-turbo=3.0.0=hcfcfb64_1
libogg=1.3.5=h2bbff1b_1
libpng=1.6.43=h19919ed_0
libsodium=1.0.18=h8d14728_1
libsqlite=3.46.0=h2466b09_0
libtiff=4.6.0=hddb2be6_3
libvorbis=1.3.7=he774522_0
libwebp-base=1.3.2=h2bbff1b_0
libxcb=1.15=hcd874cb_0
libzlib=1.3.1=h2466b09_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
matplotlib=3.8.4=py312haa95532_0
matplotlib-base=3.8.4=py312hc7c4135_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py312h2bbff1b_1
mkl_fft=1.3.8=py312h2bbff1b_0
mkl_random=1.2.4=py312h59b6b97_0
msys2-conda-epoch=20160418=1
nest-asyncio=1.6.0=pyhd8ed1ab_0
numpy=1.26.4=py312hfd52020_0
numpy-base=1.26.4=py312h4dde369_0
opencv-python=4.9.0.80=pypi_0
openjpeg=2.5.2=h3d672ee_0
openssl=3.3.1=h2466b09_0
packaging=24.0=pypi_0
parso=0.8.4=pyhd8ed1ab_0
pcre2=10.44=h3d7b363_0
pickleshare=0.7.5=py_1003
pillow=10.3.0=py312h6f6a607_0
pip=24.0=pyhd8ed1ab_0
platformdirs=4.2.2=pyhd8ed1ab_0
ply=3.11=py312haa95532_1
prompt-toolkit=3.0.46=pypi_0
psutil=5.9.8=py312he70551f_0
pthread-stubs=0.3=h3c9f919_1
pure_eval=0.2.2=pyhd8ed1ab_0
pygments=2.18.0=pyhd8ed1ab_0
pykinect-azure=0.0.3=pypi_0
pyparsing=3.0.9=py312haa95532_0
pyqt=5.15.10=py312hd77b12b_0
pyqt5-sip=12.13.0=py312h2bbff1b_0
pyqtgraph=0.13.1=py312haa95532_0
pyserial=3.5=py312haa95532_1
python=3.12.3=h2628c8c_0_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python_abi=3.12=4_cp312
pywin32=306=pypi_0
pyzmq=26.0.3=py312hd7027bb_0
qt-main=5.15.8=hcef0176_21
setuptools=70.0.0=pyhd8ed1ab_0
sip=6.7.12=py312hd77b12b_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.46.0=h2466b09_0
stack_data=0.6.2=pyhd8ed1ab_0
tbb=2021.8.0=h59b6b97_0
tk=8.6.13=h5226925_1
tornado=6.4.1=py312h4389bb4_0
traitlets=5.14.3=pyhd8ed1ab_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024a=h0c530f3_0
ucrt=10.0.22621.0=h57928b3_0
unicodedata2=15.1.0=py312h2bbff1b_0
vc=14.3=h8a93ad2_20
vc14_runtime=14.40.33810=ha82c5b3_20
vs2015_runtime=14.40.33810=h3bf8584_20
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.43.0=pyhd8ed1ab_1
xorg-libxau=1.0.11=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.4.6=h8cc25b3_1
zeromq=4.3.5=he1f189c_4
zipp=3.17.0=pypi_0
zlib=1.3.1=h2466b09_1
zstd=1.5.6=h0ea2cb4_0
